Moved code for event lists into new folder in components to be used in multiple places
