+++ src/components/EventList/components/EventItem/event-item.scss
@@ -1,0 +1,21 @@
+@import '../../../../vars';
+
+.event-item {
+  padding-left: 10px;
+  padding-top: 10px;
+  padding-right: 10px;
+  padding-bottom: 10px;
+}
+
+.event-column {
+  padding-left: 10px;
+  padding-right: 10px;
+}
+
+.event-item:hover .event-column {
+  color: #fff;
+}
+
+.event-item:hover .descriptionText {
+  color: #fff;
+}
