+++ src/components/EventList/event-list.scss
@@ -1,5 +1,27 @@
 @import '../../vars';
 
+.event-item {
+  padding-left: 10px;
+  padding-top: 10px;
+  padding-right: 10px;
+  padding-bottom: 10px;
+}
+
+.event-column {
+  padding-left: 10px;
+  padding-right: 10px;
+}
+
+.event-item:hover .event-column {
+  color: #fff;
+}
+
+.event-item:hover .descriptionText {
+  color: #fff;
+}
+
+@import '../../vars';
+
 // Make rows alternate white and grey
 /* stylelint-disable */
 .event-list > section:nth-child(even) {
